auth/integration: auth refactor init
