--- .env
@@ -1,2 +1,0 @@
-GITHUB_CLIENT_ID=
-GITHUB_CLIENT_SECRET=
