+++ src/auth/mod.rs
@@ -2,5 +2,6 @@ pub(crate) mod config;
 pub(crate) mod discord;
 pub(crate) mod github;
 pub(crate) mod gmail;
+pub(crate) mod routes;
 pub(crate) mod slack;
 pub(crate) mod telegram;
